qcom: haptic: add condition of choosing candidates for auto resonance
In case of sudden frequency drop in haptic by repetitive typing on keypad,
haptic output becomes faint and does not recover to normal by auto resonance
correction. This is attributable to the bad frequency candidates calculated
with the abnormal frequency value.
Guarantee frequency candidates are updated only when the given frequency is
in normal range(220Hz to 230Hz) and make frequency candidates be narrow
in order to prevent unsuitable frequency moving by auto resonance correction.
In addition, guarantee of auto resonance correction if running frequency is
out of range.

Bug: 23944469
Bug: 23884326

Change-Id: Iad18e372deb50542654a0bfef3ddb41ec5a5f78a
Signed-off-by: e2sh83.lee <e2sh83.lee@lge.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>